move OneHero and ListOfHeroes to separate files
